Cache directory bugs. v1.1.1

+++ src/index.ts
@@ -6,9 +6,10 @@ import Fs from "fs"
 import AvatarRoute from "./routes/avatar.js"
 import * as Constants from "./constants.js"
 
-Fs.rmSync(Constants.cacheDir, {
-  recursive: true
-});
+if(Fs.existsSync(Constants.cacheDir))
+  Fs.rmSync(Constants.cacheDir, {
+    recursive: true
+  });
 
 if(process.env.NODE_ENV !== "production") {
   Dotenv.config({ path: Path.join(process.cwd(), ".env") });
